Directory restructuring changes. Moved IB.py source to dedicated `src` folder. Changed relative imports and pyproject.toml accordingly. Moved `config.py` out of `utils` folder, now implemented as a dataclass that reads TOML instead of JSON. Moved migrations folder to the root directory for aerich, since this is unrelated to the package source.
