+++ .github/workflows/ruff.yml
@@ -6,4 +6,3 @@ jobs:
         steps:
             - uses: actions/checkout@v4
             - uses: chartboost/ruff-action@v1
-
