--- config.example.json
@@ -1,10 +1,0 @@
-{
-  "token": "",
-  "prefix": "",
-  "description": "",
-  "application_id": 0,
-  "db_host": "",
-  "db_user": "",
-  "db_name": "",
-  "db_password": ""
-}
