+++ example_config.toml
@@ -1,0 +1,8 @@
+token=""
+prefix=""
+description=""
+application_id=0
+db_host=""
+db_user=""
+db_name=""
+db_password=""
