+++ pyproject.toml
@@ -1,15 +1,3 @@
-[tool.aerich]
-tortoise_orm = "db.db.TORTOISE_ORM"
-location = "db/migrations"
-src_folder = "./."
-
-[tool.pdm]
-distribution = false
-
-[tool.pdm.dev-dependencies]
-dev = [
-    "ruff>=0.4.4",
-]
 [project]
 name = "ib.py"
 version = "0.1.0"
@@ -24,9 +12,30 @@ dependencies = [
     "aerich>=0.7.2",
     "tortoise-orm[asyncpg]>=0.20.1",
 ]
-requires-python = ">=3.11"
+requires-python = ">=3.11,<3.13"
 readme = "README.md"
 license = {text = "GPL-3.0-only"}
+
+[build-system]
+requires = ["pdm-backend"]
+build-backend = "pdm.backend"
+
+[dependency-groups]
+test = [
+    "pytest>=8.3.5",
+    "pytest-asyncio>=0.26.0",
+]
+lint = [
+    "ruff>=0.11.9",
+]
+
+[tool.pdm]
+distribution = true
+
+[tool.aerich]
+tortoise_orm = "ib_py.db.db.TORTOISE_ORM"
+location = "migrations"
+src_folder = "src/ib_py"
 
 [tool.ruff]
 line-length = 95
